moved all FCC-related registries to the FCC core package
