--- extensions/in-memory/fcc-catalogquery-adapter-registry-memory/README.md
@@ -1,4 +1,0 @@
-this is an in-memory implementation for a `ProtocolAdapterRegistry` that's used by the Federated Catalog Cache.
-
-_Note: this implementation is mainly for development purposes and might get removed/renamed/changed before the feature
-gets submitted!_
