--- extensions/in-memory/fcc-catalogquery-adapter-registry-memory/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,2 +1,0 @@
-org.eclipse.dataspaceconnector.catalog.directory.InMemoryProtocolAdapterRegistryExtension
-
