+++ settings.gradle.kts
@@ -68,8 +68,6 @@ include(":extensions:in-memory:transfer-store-memory")
 include(":extensions:in-memory:did-document-store-inmem")
 include(":extensions:in-memory:identity-hub-memory")
 include(":extensions:in-memory:fcc-node-directory-memory")
-include(":extensions:in-memory:fcc-catalogquery-adapter-registry-memory")
-include(":extensions:in-memory:fcc-query-adapter-registry-memory")
 include(":extensions:in-memory:fcc-store-memory")
 include(":extensions:iam:iam-mock")
 include(":extensions:iam:oauth2:oauth2-spi")
